feature: abstract render logic from layout component
