+++ app/components/general/header/header.css
@@ -1,3 +1,12 @@
+.header {
+  width: 100vw;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  background-color: #00150f;
+}
+
 .header-wrapper {
   width: 95vw;
   max-width: 1440px;
