+++ app/components/layout/layout.css
@@ -9,15 +9,6 @@
   align-items: center;
 }
 
-.header {
-  width: 100vw;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  background-color: #00150f;
-}
-
 .not-found-title {
   text-align: center;
 }
@@ -44,16 +35,6 @@
   display: none;
 }
 
-.layout-footer {
-  margin-top: auto;
-  height: auto;
-  width: 100vw;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
 @media screen and (max-width: 768px) {
   .container {
     width: 95vw;
@@ -62,25 +43,6 @@
     justify-content: space-between;
     align-items: center;
     max-width: 100vw;
-  }
-
-  .header {
-    width: 100vw;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-
-  .header-socials {
-    padding: 0px 10px;
-  }
-
-  .layout-footer {
-    color: white;
-    overflow-x: hidden;
-    position: relative;
-    top: 0;
-    overflow-y: scroll;
   }
 
   .overlay {
@@ -105,25 +67,6 @@
     justify-content: space-between;
     align-items: center;
     max-width: 100vw;
-  }
-
-  .header {
-    width: 100vw;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-
-  .header-socials {
-    padding: 0px 10px;
-  }
-
-  .layout-footer {
-    color: white;
-    overflow-x: hidden;
-    position: relative;
-    top: 0;
-    overflow-y: scroll;
   }
 
   .overlay {
