+++ lib/features/UI.ts
@@ -1,14 +1,10 @@
 import { createSlice } from "@reduxjs/toolkit";
 
 type InitialStateType = {
-  isMobileMenuOpen: boolean;
-  isOverlayVisible: boolean;
   loading: boolean;
 };
 
 const initialState: InitialStateType = {
-  isMobileMenuOpen: false,
-  isOverlayVisible: false,
   loading: false,
 };
 
@@ -16,17 +12,12 @@ const UISlice = createSlice({
   name: "UI",
   initialState,
   reducers: {
-    toggleMobileMenu(state, action) {
-      const { visible } = action.payload;
-      state.isMobileMenuOpen = visible;
-      state.isOverlayVisible = visible;
-    },
     setLoading(state, action) {
       state.loading = action.payload;
     },
   },
 });
 
-export const { toggleMobileMenu, setLoading } = UISlice.actions;
+export const { setLoading } = UISlice.actions;
 
 export default UISlice.reducer;
